feat: Add SVG icons for Angular, HTML5, VSCode, Spring, Git, and IntelliJ
This commit adds SVG icons for popular technologies such as Angular, HTML5, VSCode, Spring, Git, and IntelliJ. These icons can be used in the application to enhance the visual representation of these technologies.

Note: This commit message assumes that the purpose of the code changes is to add the SVG icons for the mentioned technologies. If there is any additional context or purpose behind the changes, please provide more information.