feat: Add Angular SVG icon
This commit adds the Angular SVG icon to the application. The icon enhances the visual representation of the Angular technology.

Note: This commit assumes that the purpose of the code changes is to add the Angular SVG icon. If there is any additional context or purpose behind the changes, please provide more information.