feat: Replace IntelliJ IDEA logo with Platzi logo
This commit replaces the IntelliJ IDEA logo with the Platzi logo in the README.md file. The change improves the visual representation of the technologies in the application.

Note: This commit message assumes that the purpose of the code changes is to replace the IntelliJ IDEA logo with the Platzi logo. If there is any additional context or purpose behind the changes, please provide more information.